requirements.txt
flask
PyPDF2
python-docx
pytesseract
pillow
pdf2image
pytest
pdfplumber
python-docx
opencv-python
numpy
